+++ meinberlin/apps/projects/urls.py
@@ -43,9 +43,4 @@ urlpatterns = [
         views.ProjectResultsView.as_view(),
         name="project-results",
     ),
-    re_path(
-        r"^(?P<slug>[-\w_]+)/veranstaltung/(?P<event_slug>[-\w_]+)/$",
-        views.ProjectEventView.as_view(),
-        name="project-event",
-    ),
 ]