+++ tests/offlineevents/conftest.py
@@ -2,5 +2,4 @@ from pytest_factoryboy import register
 
 from meinberlin.test.factories import offlineevents
 
-register(offlineevents.OfflineEventFactory)
 register(offlineevents.OfflineEventItemFactory)